fix(subscription): make subscriptions list thread-safe
LinkedHashSet (from mutableSetOf) is not thread-safe. Modifying it while
iterating over it causes a ConcurrentModificationException to be thrown.
This can typically happen when subscribing/unsubscribing to a
Subscription instance while notifying all subscriptions.

CopyOnWriteArraySet is useful for the case where the number of elements
is generally small and read-only operations vastly outnumber mutative
operations. This seems to match our use case here.

An alternative can be using a mutex.